# perf: using subquery in get_dr_or_cr_notes() to improve performance