[get_dr_or_cr_notes()](https://github.com/frappe/erpnext/blob/1634448864c16f4707c06e618fa963335b55184a/erpnext/accounts/doctype/payment_reconciliation/payment_reconciliation.py#L122) performs poorly as credit/debit notes count increases.

Testing on 3000 Credit notes.
<img width="1440" alt="Screenshot 2022-12-16 at 2 00 55 PM" src="https://user-images.githubusercontent.com/3272205/208058621-bf3c4d42-131e-4471-a524-995cda096056.png">

Cross Join has been replaced with Subquery.

Before:
```
In [1]: pr=frappe.get_doc('Payment Reconciliation')
   ...: pr.company='state machine'
   ...: pr.party_type='Customer'
   ...: pr.party='john doe'
   ...: pr.receivable_payable_account='Debtors - SM'

In [2]: %time res=pr.get_dr_or_cr_notes()
CPU times: user 42.4 ms, sys: 4.17 ms, total: 46.6 ms
Wall time: 2.06 s
```

After:
```
In [1]: pr=frappe.get_doc('Payment Reconciliation')
   ...: pr.company='state machine'
   ...: pr.party_type='Customer'
   ...: pr.party='john doe'
   ...: pr.receivable_payable_account='Debtors - SM'

In [2]: %time res=pr.get_dr_or_cr_notes()
CPU times: user 33.8 ms, sys: 4.2 ms, total: 38 ms
Wall time: 90.6 ms

In [3]: exit
```

`no-docs`